refactor(ui): move text-link and image-preview into primitives
