+++ packages/media/resources/js/components/detail-panel.tsx
@@ -11,7 +11,7 @@ import { Button } from "@lattice-php/ui/components/button/button";
 import { Dialog, DialogContent, DialogHeader } from "@lattice-php/ui/primitives/dialog";
 import { Input } from "@lattice-php/ui/primitives/input";
 import { Label } from "@lattice-php/ui/primitives/label";
-import { PreviewableImage } from "@lattice-php/ui/components/image/image-preview";
+import { PreviewableImage } from "@lattice-php/ui/primitives/image-preview";
 import { MODAL_MISSING_ERROR, useEmbeddedModal, useModal } from "@lattice-php/ui/modal";
 import type { MediaRow } from "./library-view";
