+++ packages/ui/resources/js/components/link/link-adapter.tsx
@@ -3,7 +3,7 @@ import type { Affix } from "@lattice-php/core";
 import { nodeIdentity } from "@lattice-php/core/test-id";
 import type { RendererComponent } from "@lattice-php/core/types";
 import { IconRenderer } from "../../icons";
-import { TextLink } from "./text-link";
+import { TextLink } from "../../primitives/text-link";
 import { actionMenuItemClassName, useActionMenu } from "../../action-menu-context";
 import { ActionTrigger, type TriggerState, useClickBehavior } from "../../click-behavior";
