+++ packages/ui/resources/js/primitives/image-preview.browser.test.tsx
