+++ packages/ui/resources/js/primitives/image-preview.tsx
@@ -1,13 +1,7 @@
-import { UI_NAMESPACE, useT } from "../../i18n";
+import { UI_NAMESPACE, useT } from "../i18n";
 import { type ComponentProps, type ReactNode, useState } from "react";
-import { Button } from "../button/button";
-import {
-  Dialog,
-  DialogClose,
-  DialogContent,
-  DialogTitle,
-  DialogTrigger,
-} from "../../primitives/dialog";
+import { Button } from "../components/button/button";
+import { Dialog, DialogClose, DialogContent, DialogTitle, DialogTrigger } from "./dialog";
 
 export type PreviewableImageProps = Omit<ComponentProps<"img">, "children"> & {
   alt: string;
